requirements.txt
appdirs==1.4.4
asgiref==3.3.4
astroid==2.2.5
autopep8==1.5.6
bleach==3.3.0
click==7.1.2
distlib==0.3.1
Django==3.1.7
django-cleanup==5.1.0
django-filter==2.4.0
django-widget-tweaks==1.4.8
filelock==3.0.12
Flask==1.1.2
future==0.18.2
gunicorn==20.1.0
html5lib==1.1
itsdangerous==1.1.0
Jinja2==2.11.3
lazy-object-proxy==1.4.1
MarkupSafe==1.1.1
numpy==1.20.2
packaging==20.9
pandas==1.2.4
Pillow==8.1.2
pycodestyle==2.7.0
pyparsing==2.4.7
PyPDF2==1.26.0
python-bidi==0.4.2
python-dateutil==2.8.1
pytz==2020.1
reportlab==3.5.66
six==1.12.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.0
virtualenv==20.4.2
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.11.2
